requirements.txt
discord.py
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Flask
pytz